only clicked bar remains highlighted in sidebar
When users click on one bar in the bar chart on the sidebarPanel, only the clicked bar (one PheWAS group) will be highlighted, the rest will grey out (in this case become more transparent). And this action is also linked to the Manhattan plot where only points in the corresponding PheWAS group will be highlighted.