Update the algorithm to calculate the comparison
Update the algorithm to calculate the comparison of encounters between two consecutively-recorded years/months/days. 
Now: 1) As long as there are some certain encounter types this year, then even if there is no such type in the previous year, set the corresponding type in the previous year to zero.   - [actually this change has mostly been made in the last commit]
2) ... more to follow up